% Auto-generated: do not edit by hand
\name{gridMap}

\alias{gridMap}

\title{GridMap component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
gridMap(id=NULL, className=NULL, data=NULL, height=NULL,
show_axes=NULL, style=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. Classname .}

\item{data}{List of named lists. Dataset.}

\item{height}{Numeric. The value displayed in the input.}

\item{show_axes}{Logical. Dash-assigned callback that should be called to report property changes
to Dash, to make them available for callbacks.}

\item{style}{Named list. The value displayed in the input.}

\item{width}{Numeric. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

